Same Year Correction
If there was a birth and death in the same year, the interviewer may say that it counts as a full year (cause even if it's a one day difference, it could be the peak day in all the days between the first birth and last death). So we will count this year and decrement the year after.